// src/ContentDashboard/Client/src/api/types.gen.ts
export type ContentOwnerResponseModel = {
    id: string;
    name: string;
    email?: null | string;
    documentCount: number;
};

export type DocumentOwnerModel = {
    id?: null | string;
    name: string;
};

export type OwnedDocumentResponseModel = {
    id: string;
    name: string;
    contentTypeAlias: string;
    icon?: null | string;
    published: boolean;
    edited: boolean;
    updateDate: string;
    createDate: string;
    owner?: null | DocumentOwnerModel;
};

export type OwnedDocumentSortField = 'Name' | 'UpdateDate';
// ...
export type SortDirection = 'Ascending' | 'Descending';